+++ docs/service-worker.js
@@ -14,7 +14,7 @@
 importScripts("https://storage.googleapis.com/workbox-cdn/releases/4.3.1/workbox-sw.js");
 
 importScripts(
-  "./precache-manifest.653689b5f6bf1ab18fa32ef8ab30a142.js"
+  "./precache-manifest.97713dc0ffe7b7f49aa356ced8e9e402.js"
 );
 
 self.addEventListener('message', (event) => {