+++ src/components/Sidebar/LinksGroup/LinksGroup.js
@@ -234,10 +234,4 @@ class LinksGroup extends Component {
   }
 }
 
-function mapStateToProps(store) {
-  return {
-    sidebarColor: store.layout.sidebarColor
-  };
-}
-
-export default withRouter(connect(mapStateToProps)(LinksGroup));
+export default LinksGroup;